+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -1,12 +1,11 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.service.FilmService;
+import ru.yandex.practicum.filmorate.service.impl.FilmServiceImpl;
 import jakarta.validation.Valid;
 
 import java.util.List;
@@ -16,24 +15,23 @@ import java.util.List;
 @RestController
 @RequestMapping("/films")
 public class FilmController {
-    private final FilmService filmService;
+    private final FilmServiceImpl filmServiceImpl;
 
-    @Autowired
-    public FilmController(FilmService filmService) {
-        this.filmService = filmService;
+    public FilmController(FilmServiceImpl filmServiceImpl) {
+        this.filmServiceImpl = filmServiceImpl;
     }
 
     @GetMapping
     public ResponseEntity<List<Film>> getAllFilms() {
         log.info("Получен запрос на получение всех фильмов");
-        List<Film> films = filmService.getAllFilms();
+        List<Film> films = filmServiceImpl.getAllFilms();
         return ResponseEntity.ok(films);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Film> getFilm(@PathVariable int id) {
         log.info("Получен запрос на получение фильма с ID {}", id);
-        Film film = filmService.getFilm(id);
+        Film film = filmServiceImpl.getFilm(id);
         return ResponseEntity.ok(film);
     }
 
@@ -41,7 +39,7 @@ public class FilmController {
     public ResponseEntity<Film> createFilm(@Valid @RequestBody Film film) {
         log.info("Получен запрос на создание фильма: {}", film);
         film.validate();
-        Film createdFilm = filmService.addFilm(film);
+        Film createdFilm = filmServiceImpl.addFilm(film);
         return ResponseEntity.ok(createdFilm);
     }
 
@@ -49,28 +47,28 @@ public class FilmController {
     public ResponseEntity<Film> updateFilm(@Valid @RequestBody Film film) {
         log.info("Получен запрос на обновление фильма с ID {}", film.getId());
         film.validate();
-        Film updatedFilm = filmService.updateFilm(film);
+        Film updatedFilm = filmServiceImpl.updateFilm(film);
         return ResponseEntity.ok(updatedFilm);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteFilm(@PathVariable int id) {
         log.info("Получен запрос на удаление фильма с ID {}", id);
-        filmService.deleteFilm(id);
+        filmServiceImpl.deleteFilm(id);
         return ResponseEntity.noContent().build();
     }
 
     @PutMapping("/{id}/like/{userId}")
     public ResponseEntity<Void> addLike(@PathVariable int id, @PathVariable int userId) {
         log.info("Получен запрос на добавление лайка фильму {} от пользователя {}", id, userId);
-        filmService.addLike(id, userId);
+        filmServiceImpl.addLike(id, userId);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}/like/{userId}")
     public ResponseEntity<Void> removeLike(@PathVariable int id, @PathVariable int userId) {
         log.info("Получен запрос на удаление лайка у фильма {} от пользователя {}", id, userId);
-        filmService.removeLike(id, userId);
+        filmServiceImpl.removeLike(id, userId);
         return ResponseEntity.noContent().build();
     }
 
@@ -78,7 +76,7 @@ public class FilmController {
     public ResponseEntity<List<Film>> getPopularFilms(
             @RequestParam(defaultValue = "10") int count) {
         log.info("Получен запрос на получение {} популярных фильмов", count);
-        List<Film> popularFilms = filmService.getPopularFilms(count);
+        List<Film> popularFilms = filmServiceImpl.getPopularFilms(count);
         return ResponseEntity.ok(popularFilms);
     }
 }
