+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -5,7 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.service.UserService;
+import ru.yandex.practicum.filmorate.service.impl.UserServiceImpl;
 import jakarta.validation.Valid;
 
 import java.util.List;
@@ -15,66 +15,66 @@ import java.util.List;
 @RestController
 @RequestMapping("/users")
 public class UserController {
-    private final UserService userService;
+    private final UserServiceImpl userServiceImpl;
 
     @Autowired
-    public UserController(UserService userService) {
-        this.userService = userService;
+    public UserController(UserServiceImpl userServiceImpl) {
+        this.userServiceImpl = userServiceImpl;
     }
 
     @PostMapping
     public User createUser(@Valid @RequestBody User user) {
         log.info("Получен запрос на создание пользователя: {}", user);
         user.prepareForSave();
-        return userService.createUser(user);
+        return userServiceImpl.createUser(user);
     }
 
     @PutMapping
     public User updateUser(@Valid @RequestBody User user) {
         log.info("Получен запрос на обновление пользователя с ID {}: {}", user.getId(), user);
         user.prepareForSave();
-        return userService.updateUser(user);
+        return userServiceImpl.updateUser(user);
     }
 
     @GetMapping
     public List<User> getAllUsers() {
         log.info("Получен запрос на получение всех пользователей");
-        return userService.getAllUsers();
+        return userServiceImpl.getAllUsers();
     }
 
     @GetMapping("/{id}")
     public User getUser(@PathVariable int id) {
         log.info("Получен запрос на получение пользователя с ID {}", id);
-        return userService.getUser(id);
+        return userServiceImpl.getUser(id);
     }
 
     @DeleteMapping("/{id}")
     public void deleteUser(@PathVariable int id) {
         log.info("Получен запрос на удаление пользователя с ID {}", id);
-        userService.deleteUser(id);
+        userServiceImpl.deleteUser(id);
     }
 
     @PutMapping("/{id}/friends/{friendId}")
     public void addFriend(@PathVariable int id, @PathVariable int friendId) {
         log.info("Получен запрос на добавление в друзья от пользователя {} пользователю {}", id, friendId);
-        userService.addFriend(id, friendId);
+        userServiceImpl.addFriend(id, friendId);
     }
 
     @DeleteMapping("/{id}/friends/{friendId}")
     public void removeFriend(@PathVariable int id, @PathVariable int friendId) {
         log.info("Получен запрос на удаление из друзей пользователя {} у пользователя {}", friendId, id);
-        userService.removeFriend(id, friendId);
+        userServiceImpl.removeFriend(id, friendId);
     }
 
     @GetMapping("/{id}/friends")
     public List<User> getFriends(@PathVariable int id) {
         log.info("Получен запрос на получение списка друзей пользователя с ID {}", id);
-        return userService.getFriends(id);
+        return userServiceImpl.getFriends(id);
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
     public List<User> getCommonFriends(@PathVariable int id, @PathVariable int otherId) {
         log.info("Получен запрос на получение общих друзей пользователей {} и {}", id, otherId);
-        return userService.getCommonFriends(id, otherId);
+        return userServiceImpl.getCommonFriends(id, otherId);
     }
 }
