+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,84 +1,23 @@
 package ru.yandex.practicum.filmorate.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.exception.ResourceNotFoundException;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
 
-import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-public class FilmService {
-    private final FilmStorage filmStorage;
-    private final UserService userService;
+public interface FilmService {
+    Film addFilm(Film film);
 
-    @Autowired
-    public FilmService(FilmStorage filmStorage, UserService userService) {
-        this.filmStorage = filmStorage;
-        this.userService = userService;
-    }
+    Film updateFilm(Film film);
 
-    public Film addFilm(Film film) {
-        return filmStorage.addFilm(film);
-    }
+    Film getFilm(int id);
 
-    public Film updateFilm(Film film) {
-        Film updatedFilm = filmStorage.updateFilm(film);
-        if (updatedFilm == null) {
-            throw new ResourceNotFoundException("Фильм с ID " + film.getId() + " не найден");
-        }
-        return updatedFilm;
-    }
+    List<Film> getAllFilms();
 
-    public Film getFilm(int id) {
-        Film film = filmStorage.getFilm(id);
-        if (film == null) {
-            throw new ResourceNotFoundException("Фильм с ID " + id + " не найден");
-        }
-        return film;
-    }
+    void deleteFilm(int id);
 
-    public List<Film> getAllFilms() {
-        return filmStorage.getAllFilms();
-    }
+    void addLike(int filmId, int userId);
 
-    public void deleteFilm(int id) {
-        filmStorage.deleteFilm(id);
-    }
+    void removeLike(int filmId, int userId);
 
-    public void addLike(int filmId, int userId) {
-        Film film = getFilm(filmId);
-        userService.getUser(userId);
-
-        if (film.getLikes().contains(userId)) {
-            throw new ValidationException("Пользователь уже поставил лайк этому фильму");
-        }
-        film.getLikes().add(userId);
-    }
-
-    public void removeLike(int filmId, int userId) {
-        Film film = getFilm(filmId);
-        if (!film.getLikes().remove(userId)) {
-            throw new ResourceNotFoundException("Лайк от пользователя " + userId + " не найден");
-        }
-    }
-
-    public List<Film> getPopularFilms(int count) {
-        return filmStorage.getAllFilms().stream()
-                .sorted(Comparator.comparingInt(f -> -f.getLikes().size()))
-                .limit(count)
-                .collect(Collectors.toList());
-    }
-
-    private Film getFilmOrThrow(int id) {
-        Film film = filmStorage.getFilm(id);
-        if (film == null) {
-            throw new ResourceNotFoundException("Фильм с ID " + id + " не найден");
-        }
-        return film;
-    }
+    List<Film> getPopularFilms(int count);
 }
