+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,97 +1,25 @@
 package ru.yandex.practicum.filmorate.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.exception.ResourceNotFoundException;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-public class UserService {
-    private final UserStorage userStorage;
+public interface UserService {
+    User createUser(User user);
 
-    @Autowired
-    public UserService(UserStorage userStorage) {
-        this.userStorage = userStorage;
-    }
+    User updateUser(User user);
 
-    public User createUser(User user) {
-        return userStorage.createUser(user);
-    }
+    User getUser(int id);
 
-    public User updateUser(User user) {
-        User updatedUser = userStorage.updateUser(user);
-        if (updatedUser == null) {
-            throw new ResourceNotFoundException("Пользователь с ID " + user.getId() + " не найден");
-        }
-        return updatedUser;
-    }
+    List<User> getAllUsers();
 
-    public User getUser(int id) {
-        User user = userStorage.getUser(id);
-        if (user == null) {
-            throw new ResourceNotFoundException("Пользователь с ID " + id + " не найден");
-        }
-        return user;
-    }
+    void deleteUser(int id);
 
-    public User getUserOrThrow(int id) {
-        User user = userStorage.getUser(id);
-        if (user == null) {
-            throw new ResourceNotFoundException("Пользователь с ID " + id + " не найден");
-        }
-        return user;
-    }
+    void addFriend(int userId, int friendId);
 
-    public List<User> getAllUsers() {
-        return userStorage.getAllUsers();
-    }
+    void removeFriend(int userId, int friendId);
 
-    public void deleteUser(int id) {
-        if (userStorage.getUser(id) == null) {
-            throw new ResourceNotFoundException("Пользователь с ID " + id + " не найден");
-        }
-        userStorage.deleteUser(id);
-    }
+    List<User> getFriends(int userId);
 
-    public void addFriend(int userId, int friendId) {
-        User user = getUser(userId);
-        User friend = getUser(friendId);
-
-        if (user.getFriends().contains(friendId)) {
-            throw new ValidationException("Пользователи уже являются друзьями");
-        }
-
-        user.getFriends().add(friendId);
-        friend.getFriends().add(userId);
-    }
-
-    public void removeFriend(int userId, int friendId) {
-        User user = getUser(userId);
-        User friend = getUser(friendId);
-
-        user.getFriends().remove(friendId);
-        friend.getFriends().remove(userId);
-    }
-
-    public List<User> getFriends(int userId) {
-        User user = getUser(userId);
-        return user.getFriends().stream()
-                .map(this::getUser)
-                .collect(Collectors.toList());
-    }
-
-    public List<User> getCommonFriends(int userId, int otherUserId) {
-        User user = getUser(userId);
-        User otherUser = getUser(otherUserId);
-
-        return user.getFriends().stream()
-                .filter(otherUser.getFriends()::contains)
-                .map(this::getUser)
-                .collect(Collectors.toList());
-    }
+    List<User> getCommonFriends(int userId, int otherUserId);
 }
